# docs/openapi.yaml
openapi: 3.1.0
info:
  description: |-
    The PaymentFlow payments API. This document describes the public `/v1` surface — the endpoints PaymentFlow versions, deprecates on a published timeline, and supports for integrators. Endpoints not described here are internal and carry no compatibility promise.

    Amounts are integers in the currency's minor unit (`amountMinor`): `1000` in `USD` is $10.00. There are no floating-point amounts anywhere in this API.

    Every mutating request accepts an `Idempotency-Key` header and every list is cursor-paginated; both are described on the operations that use them.
  license:
    name: Proprietary
  title: PaymentFlow API
  version: 2026-08-01
servers:
- url: https://api.paymentflow.dev
  description: PaymentFlow API
security:
- SecretKey: []
paths:
...
  /v1/request_logs:
    get:
      description: |-
        Returns the API calls you have made, most recent first, cursor-paginated — the record you reach for when an integration behaved differently from how you expected.

        Each entry carries the `requestId` the response returned, so an error a customer reported can be found here by the identifier they quote. Bodies and headers are **redacted at the edge before they are stored**: card numbers, API keys and secrets never reach this log, so what you see here is what was safe to keep rather than what was sent.

        Entries are pruned on a schedule; older traffic survives as aggregates at `/v1/usage`.
      operationId: listRequestLogs
      parameters:
      - description: "How many objects to return, 1–100. Values above the ceiling are clamped to it rather than rejected, so a client asking for more simply receives the maximum. Defaults to 25."
        in: query
        name: limit
        required: false
        schema:
          type: integer
          format: int32
      - description: A cursor for pagination. Pass the `nextCursor` from the previous response to fetch the following page. The value is opaque and signed — treat it as a token and never parse or construct one.
        in: query
        name: starting_after
        required: false
        schema:
          type: string
      - description: "Return only objects created at or after this instant, as RFC 3339 (`2026-07-29T00:00:00Z`)."
        in: query
        name: created_after
        required: false
        schema:
          type: string
          format: date-time
      - description: "Return only objects created at or before this instant, as RFC 3339 (`2026-07-29T23:59:59Z`)."
        in: query
        name: created_before
        required: false
        schema:
          type: string
          format: date-time
      - description: Return only requests that returned this exact HTTP status.
        example: 409
        in: query
        name: status_code
        required: false
        schema:
          type: integer
          format: int32
      - description: Return only requests made with this HTTP method.
        example: POST
        in: query
        name: method
        required: false
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CursorPageRequestLogResponse"
          description: "A page of request log entries, most recent first."
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
          description: "The query could not be understood: a tampered or malformed `starting_after` cursor, a `limit` that is not positive, or a filter value outside the vocabulary this endpoint accepts. A rejected filter is deliberately an error rather than an empty page — an empty page is something you would then have to explain to yourself."
        "401":
          content:
            application/json:
              example:
                code: UNAUTHORIZED
                correlationId: 3f9a1c2e-77b4-4a1d-9c3e-1d2b4f6a8c05
                docUrl: https://docs.paymentflow.dev/errors#unauthorized
                message: Authentication is required or has failed.
                path: /v1/payments
                requestId: req_9f2c1e7a4b8d
                status: 401
                timestamp: 2026-07-28T12:34:56.789Z
                type: authentication_error
              schema:
                $ref: "#/components/schemas/ApiError"
          description: Authentication is required or has failed.
        "403":
          content:
            application/json:
              example:
                code: INSUFFICIENT_SCOPE
                correlationId: 3f9a1c2e-77b4-4a1d-9c3e-1d2b4f6a8c05
                docUrl: https://docs.paymentflow.dev/errors#insufficient_scope
                message: This API key does not have the required scope for this endpoint.
                path: /v1/payments
                requestId: req_9f2c1e7a4b8d
                status: 403
                timestamp: 2026-07-28T12:34:56.789Z
                type: permission_error
              schema:
                $ref: "#/components/schemas/ApiError"
          description: This API key does not have the required scope for this endpoint.
        "429":
          content:
            application/json:
              example:
                code: RATE_LIMIT_EXCEEDED
                correlationId: 3f9a1c2e-77b4-4a1d-9c3e-1d2b4f6a8c05
                docUrl: https://docs.paymentflow.dev/errors#rate_limit_exceeded
                message: Too many requests. Slow down and retry after the interval in the Retry-After header.
                path: /v1/payments
                requestId: req_9f2c1e7a4b8d
                status: 429
                timestamp: 2026-07-28T12:34:56.789Z
                type: rate_limit_error
              schema:
                $ref: "#/components/schemas/ApiError"
          description: Too many requests. Slow down and retry after the interval in the Retry-After header.
        "500":
          content:
            application/json:
              example:
                code: INTERNAL_ERROR
                correlationId: 3f9a1c2e-77b4-4a1d-9c3e-1d2b4f6a8c05
                docUrl: https://docs.paymentflow.dev/errors#internal_error
                message: An unexpected error occurred.
                path: /v1/payments
                requestId: req_9f2c1e7a4b8d
                status: 500
                timestamp: 2026-07-28T12:34:56.789Z
                type: api_error
              schema:
                $ref: "#/components/schemas/ApiError"
          description: An unexpected error occurred.
      summary: List API request logs
      tags:
      - Request logs
...
  /v1/usage:
    get:
      description: |-
        Returns how much of the API you used over a date range, broken down by day, key and route, with latency percentiles per bucket.

        Dated rather than timestamped, because the aggregate is per UTC day — accepting an instant would imply a precision the stored data does not have. Usage is grouped by key, so two keys hitting the same route on the same day appear as two buckets rather than one merged number; percentiles cannot be meaningfully averaged across keys, which is why they are not.

        This outlives the raw request log: entries there are pruned, and their counts are rolled forward into these aggregates first.
      operationId: getUsage
      parameters:
      - description: "The first day to include, in UTC. Defaults to 30 days ago."
        example: 2026-07-01
        in: query
        name: from
        required: false
        schema:
          type: string
          format: date
      - description: "The last day to include, in UTC. Defaults to today."
        example: 2026-07-29
        in: query
        name: to
        required: false
        schema:
          type: string
          format: date
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsageSummaryResponse"
          description: "Usage totals for the range, and the per-day, per-key, per-route breakdown."
        "401":
          content:
            application/json:
              example:
                code: UNAUTHORIZED
                correlationId: 3f9a1c2e-77b4-4a1d-9c3e-1d2b4f6a8c05
                docUrl: https://docs.paymentflow.dev/errors#unauthorized
                message: Authentication is required or has failed.
                path: /v1/payments
                requestId: req_9f2c1e7a4b8d
                status: 401
                timestamp: 2026-07-28T12:34:56.789Z
                type: authentication_error
              schema:
                $ref: "#/components/schemas/ApiError"
          description: Authentication is required or has failed.
        "403":
          content:
            application/json:
              example:
                code: INSUFFICIENT_SCOPE
                correlationId: 3f9a1c2e-77b4-4a1d-9c3e-1d2b4f6a8c05
                docUrl: https://docs.paymentflow.dev/errors#insufficient_scope
                message: This API key does not have the required scope for this endpoint.
                path: /v1/payments
                requestId: req_9f2c1e7a4b8d
                status: 403
                timestamp: 2026-07-28T12:34:56.789Z
                type: permission_error
              schema:
                $ref: "#/components/schemas/ApiError"
          description: This API key does not have the required scope for this endpoint.
        "429":
          content:
            application/json:
              example:
                code: RATE_LIMIT_EXCEEDED
                correlationId: 3f9a1c2e-77b4-4a1d-9c3e-1d2b4f6a8c05
                docUrl: https://docs.paymentflow.dev/errors#rate_limit_exceeded
                message: Too many requests. Slow down and retry after the interval in the Retry-After header.
                path: /v1/payments
                requestId: req_9f2c1e7a4b8d
                status: 429
                timestamp: 2026-07-28T12:34:56.789Z
                type: rate_limit_error
              schema:
                $ref: "#/components/schemas/ApiError"
          description: Too many requests. Slow down and retry after the interval in the Retry-After header.
        "500":
          content:
            application/json:
              example:
                code: INTERNAL_ERROR
                correlationId: 3f9a1c2e-77b4-4a1d-9c3e-1d2b4f6a8c05
                docUrl: https://docs.paymentflow.dev/errors#internal_error
                message: An unexpected error occurred.
                path: /v1/payments
                requestId: req_9f2c1e7a4b8d
                status: 500
                timestamp: 2026-07-28T12:34:56.789Z
                type: api_error
              schema:
                $ref: "#/components/schemas/ApiError"
          description: An unexpected error occurred.
      summary: Summarize API usage
      tags:
      - Usage
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        code:
          type: string
          description: "A stable, machine-readable identifier for this specific failure, such as `PAYMENT_NOT_CAPTURABLE`. The set of codes grows by policy, which is why `type` and not this is what a client should switch on. Every code is catalogued in docs/ERRORS.md."
        correlationId:
          type: string
          description: "Identifies the whole distributed trace this call belongs to, which may span several services. Broader than `requestId`."
        docUrl:
          type: string
          description: Where to read about this specific code.
        errors:
          type: array
          description: "Field-level validation failures, when the request failed validation in more than one place."
          items:
            $ref: "#/components/schemas/ApiFieldError"
        message:
          type: string
          description: "A human-readable explanation. Safe to log; not intended to be shown to your customers, and never programmatically parsed."
        param:
          type: string
          description: "The single offending parameter, when there is exactly one. Multi-field validation failures use `errors` instead."
        path:
          type: string
          description: The request path that produced the error.
        requestId:
          type: string
          description: Identifies this one HTTP call. **Quote this in a support request** — it is on every log line and every row of your request log.
        status:
          type: integer
          format: int32
          description: "The HTTP status code, repeated in the body so a logged response is self-contained."
        timestamp:
          type: string
          format: date-time
          description: "When the error occurred, as RFC 3339."
        type:
          type: string
          description: "The error's classification, and **the field to branch on**. This is a small closed set an SDK maps to exception classes: `authentication_error`, `permission_error`, `invalid_request_error`, `idempotency_error`, `rate_limit_error`, `api_error`."
    ApiFieldError:
      type: object
      properties:
        field:
          type: string
          description: The request field that failed validation.
        message:
          type: string
          description: Why it failed. The rejected value is deliberately never echoed back.
...
    CursorPageRequestLogResponse:
      type: object
      properties:
        data:
          type: array
          description: "The objects on this page, most recent first."
          items:
            $ref: "#/components/schemas/RequestLogResponse"
        hasMore:
          type: boolean
          description: Whether more objects exist after this page. A cursor page reports no total count — that would cost a second full count query on every request — and this is the only question a paginating client needs.
        nextCursor:
          type: string
          description: "Pass this as `starting_after` to fetch the next page. Absent when `hasMore` is false. Opaque and signed: treat it as a token, never parse or construct one."
        object:
          type: string
          description: "Always `list`. A constant discriminator, so a client deserializing a response can branch on a field rather than on the endpoint it called."
...
    RequestLogResponse:
      type: object
      properties:
        clientIp:
          type: string
          description: The client address the request arrived from.
          example: 203.0.113.7
        correlationId:
          type: string
          description: Identifies the whole distributed trace this call belongs to.
        durationMs:
          type: integer
          format: int64
          description: "How long the platform took to answer, in milliseconds. Server time only — it excludes the network."
          example: 42
        errorCode:
          type: string
          description: "The catalogued error code, when the request failed. Absent for a successful request."
          example: PAYMENT_NOT_CAPTURABLE
        id:
          type: string
          format: uuid
          description: Unique identifier for this log entry.
        keyId:
          type: string
          format: uuid
          description: The API key the request was made with. Absent for a key that has since been deleted.
        method:
          type: string
          description: The HTTP method.
          example: POST
        mode:
          type: string
          description: Whether the request was made with a `test` or `live` key.
          enum:
          - test
          - live
          example: test
        object:
          type: string
          description: Always `request_log`.
          example: request_log
        occurredAt:
          type: string
          format: date-time
          description: "When the request arrived, as RFC 3339."
        path:
          type: string
          description: "The path requested, as sent."
          example: /v1/payments
        queryString:
          type: string
          description: "The query string, if any, with sensitive values already redacted."
          example: limit=25
        requestBody:
          type: string
          description: "The request body, **already redacted and truncated** at the edge before it was ever stored — card numbers, keys and secrets never reach this log."
        requestHeaders:
          type: object
          additionalProperties:
            type: string
          description: "The request headers, with credential-bearing values redacted."
        requestId:
          type: string
          description: "Identifies this one HTTP call — the same value the response's error body carried, if it failed."
          example: req_9f2c1e7a4b8d
        responseBody:
          type: string
          description: "The response body, redacted and truncated on the same terms as the request body."
        statusCode:
          type: integer
          format: int32
          description: The HTTP status returned.
          example: 201
        userAgent:
          type: string
          description: "The `User-Agent` header, useful for telling one of your services from another."
          example: paymentflow-node/1.2.0
...
    UsageBucketResponse:
      type: object
      properties:
        clientErrors:
          type: integer
          format: int64
          description: How many returned a 4xx.
          example: 11
        day:
          type: string
          format: date
          description: "The day this bucket covers, in UTC."
          example: 2026-07-29
        keyId:
          type:
          - string
          - "null"
          format: uuid
          description: "The API key the traffic was made with. Usage is grouped by key, so two keys hitting the same route on the same day produce two buckets. **Null** for traffic made with a key that has since been deleted — the usage is still a fact about your day."
        maxDurationMs:
          type: integer
          format: int64
          description: "The slowest single request in the bucket, in milliseconds."
          example: 812
        meanDurationMs:
          type:
          - integer
          - "null"
          format: int64
          description: Mean server-side duration in milliseconds. Null when the bucket had no traffic.
          example: 42
        p50DurationMs:
          type:
          - integer
          - "null"
          format: int64
          description: Median server-side duration in milliseconds. Null when the bucket had no traffic.
          example: 31
        p95DurationMs:
          type:
          - integer
          - "null"
          format: int64
          description: "95th-percentile server-side duration in milliseconds. **Explicitly null**, never omitted, when the bucket had no traffic: a percentile over zero requests is unknown rather than zero."
          example: 118
        p99DurationMs:
          type:
          - integer
          - "null"
          format: int64
          description: 99th-percentile server-side duration in milliseconds. Null when the bucket had no traffic.
          example: 260
        requests:
          type: integer
          format: int64
          description: How many requests fell in this bucket.
          example: 4820
        route:
          type: string
          description: "The route pattern the requests hit, not the concrete path — so every payment retrieval aggregates together."
          example: "/v1/payments/{id}"
        serverErrors:
          type: integer
          format: int64
          description: How many returned a 5xx.
          example: 0
    UsageSummaryResponse:
      type: object
      properties:
        buckets:
          type: array
          description: "The breakdown behind the totals, one bucket per day, key and route."
          items:
            $ref: "#/components/schemas/UsageBucketResponse"
        from:
          type: string
          format: date
          description: "The first day these totals cover, inclusive."
          example: 2026-07-01
        to:
          type: string
          format: date
          description: "The last day these totals cover, inclusive."
          example: 2026-07-29
        totalClientErrors:
          type: integer
          format: int64
          description: How many of those returned a 4xx — your requests that were rejected.
          example: 312
        totalRequests:
          type: integer
          format: int64
          description: How many API requests you made across the range.
          example: 148203
        totalServerErrors:
          type: integer
          format: int64
          description: "How many returned a 5xx — this platform's failures, not yours."
          example: 4
...
  securitySchemes:
    SecretKey:
      bearerFormat: sk
      description: |-
        Your secret API key, sent as `Authorization: Bearer sk_test_...`.

        The key determines both *whose* data you see and *which mode* you see it in: an `sk_test_` key reads and writes test data only, an `sk_live_` key live data only. No header, parameter, or body field can change either — to work in the other mode, use the other key.
      scheme: bearer
      type: http